x/warp: move warp msg handler to separate file
